% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{.baseline_drop}
\alias{.baseline_drop}
\title{Drop pins from a board's baseline}
\usage{
.baseline_drop(x, name)
}
\arguments{
\item{x}{A \code{gdpins_board}.}

\item{name}{Character vector of pin names.}
}
\description{
Drop pins from a board's baseline
}
\keyword{internal}
