% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{.baseline_read}
\alias{.baseline_read}
\title{Read a board's baseline}
\usage{
.baseline_read(x, quiet = FALSE)
}
\arguments{
\item{x}{A \code{gdpins_board}.}

\item{quiet}{Passed to \code{.baseline_read_file()}.}
}
\value{
A data frame or \code{NULL}.
}
\description{
Read a board's baseline
}
\keyword{internal}
