% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{.collect_tables_work}
\alias{.collect_tables_work}
\title{Collect table publish work items from a gdpins_board}
\usage{
.collect_tables_work(tables_board)
}
\arguments{
\item{tables_board}{A \code{gdpins_board} or \code{NULL}.}
}
\value{
A tibble with column \code{pin_name}.
}
\description{
Collect table publish work items from a gdpins_board
}
\keyword{internal}
