% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{.empty_prune_plan}
\alias{.empty_prune_plan}
\title{An empty prune plan}
\usage{
.empty_prune_plan()
}
\value{
A 0-row tibble with the prune plan columns \code{name}, \code{version},
\code{side}, \code{hash} and \code{action}.
}
\description{
An empty prune plan
}
\keyword{internal}
