% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{.raw_baseline_drop}
\alias{.raw_baseline_drop}
\title{Drop raw files from a connection's baseline}
\usage{
.raw_baseline_drop(conn, name)
}
\description{
Drop raw files from a connection's baseline
}
\keyword{internal}
