% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive-adapter.R
\name{.real_hits_to_tbl}
\alias{.real_hits_to_tbl}
\title{Convert a drive_ls dribble to the standard adapter tibble with a path prefix}
\usage{
.real_hits_to_tbl(hits, prefix = "")
}
\description{
Convert a drive_ls dribble to the standard adapter tibble with a path prefix
}
\keyword{internal}
