% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{.resolve_on_discrepancy}
\alias{.resolve_on_discrepancy}
\title{Resolve the default on_discrepancy value}
\usage{
.resolve_on_discrepancy(on_discrepancy)
}
\description{
Resolve the default on_discrepancy value
}
\keyword{internal}
