% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{.write_pin_snapshot}
\alias{.write_pin_snapshot}
\title{Write a pin snapshot to a board}
\usage{
.write_pin_snapshot(dst_board, snap, pin_name)
}
\arguments{
\item{dst_board}{A pins board.}

\item{snap}{A snapshot from \code{.snapshot_pin()}.}

\item{pin_name}{Character scalar.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Write a pin snapshot to a board
}
\keyword{internal}
