% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{.read_from_board}
\alias{.read_from_board}
\title{Read from a single pins board}
\usage{
.read_from_board(pins_board, name, version)
}
\arguments{
\item{pins_board}{A pins board object.}

\item{name}{Pin name.}

\item{version}{Character scalar or \code{NULL}.}
}
\description{
Read from a single pins board
}
\keyword{internal}
