% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{.write_to_board}
\alias{.write_to_board}
\title{Write x to a single pins board with appropriate type}
\usage{
.write_to_board(pins_board, x, name, fmt, versioned)
}
\arguments{
\item{pins_board}{A pins board object.}

\item{x}{The R object to write.}

\item{name}{Pin name.}

\item{fmt}{Character scalar: \code{"parquet"} or \code{"rds"}.}

\item{versioned}{Logical. Whether this write creates a version.}
}
\description{
Write x to a single pins board with appropriate type
}
\keyword{internal}
