% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive-adapter.R
\name{gd_exists}
\alias{gd_exists}
\title{Check if a path exists on Drive}
\usage{
gd_exists(adapter, path)
}
\arguments{
\item{adapter}{A \code{gdpins_drive_adapter} object.}

\item{path}{Character scalar. Path relative to the adapter root.}
}
\value{
Logical scalar.
}
\description{
Check if a path exists on Drive
}
\keyword{internal}
