% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-connection.R
\name{gdpins_raw_ls}
\alias{gdpins_raw_ls}
\title{List contents of a raw connection}
\usage{
gdpins_raw_ls(conn, depth = 2)
}
\arguments{
\item{conn}{A \code{gdpins_raw_conn} object.}

\item{depth}{Integer scalar. Maximum directory depth to display. Default \code{2}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} describing the folder tree.
}
\description{
Returns a compact folder tree for the raw connection, defaulting to 2 levels
of depth.
}
