% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-connection.R
\name{gdpins_refresh_disconnect}
\alias{gdpins_refresh_disconnect}
\title{Force-refresh a raw connection and disconnect}
\usage{
gdpins_refresh_disconnect(conn)
}
\arguments{
\item{conn}{A \code{gdpins_raw_conn} object.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Pulls all files from Drive to the local mirror, then invalidates the
connection object.
}
\seealso{
Other raw-connection: 
\code{\link{gdpins_raw_connect}()},
\code{\link{gdpins_raw_get}()},
\code{\link{gdpins_raw_ls}()},
\code{\link{gdpins_raw_path}()},
\code{\link{gdpins_raw_put_file}()},
\code{\link{gdpins_raw_put_object}()},
\code{\link{gdpins_raw_remove}()}
}
\concept{raw-connection}
