% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{print.gdpins_board}
\alias{print.gdpins_board}
\title{Print a gdpins_board object (compact, ≤80 cols)}
\usage{
\method{print}{gdpins_board}(x, ...)
}
\arguments{
\item{x}{A \code{gdpins_board} object.}

\item{...}{Unused.}
}
\value{
Invisibly \code{x}.
}
\description{
Print a gdpins_board object (compact, ≤80 cols)
}
